+++ src/main/kotlin/jmessenger/jlanguage/messages/Attachment.kt
@@ -1,6 +1,6 @@
 package jmessenger.jlanguage.messages
 
-open class Attachment(type: Short, var id: Int = 0) : JMessage(type) {
+open class Attachment(type: Int, var id: Int = 0) : JMessage(type) {
 
     override fun equals(other: Any?) = if (other is Attachment) id == other.id
     else false
