+++ src/main/kotlin/jmessenger/jlanguage/messages/JMessage.kt
@@ -2,4 +2,4 @@ package jmessenger.jlanguage.messages
 
 import jmessenger.jlanguage.Ignore
 
-open class JMessage(@Ignore var type: Short)
+open class JMessage(@Ignore var type: Int)
