+++ src/main/kotlin/jmessenger/jlanguage/messages/requests/RequestMessage.kt
@@ -2,4 +2,4 @@ package jmessenger.jlanguage.messages.requests
 
 import jmessenger.jlanguage.messages.JMessage
 
-open class RequestMessage(type: Short, var requestId: Int = 0) : JMessage(type)
+open class RequestMessage(type: Int, var requestId: Int = 0) : JMessage(type)
